Move LSMStore to its own file and switch to threadsafe RBTree impl
