+++ src/store/mod.rs
@@ -1,9 +1,9 @@
+pub mod lsm;
+
 use self::TransactInstruction::{Delete, Set};
 use crate::Result;
 use async_trait::async_trait;
-use bloom_filter_rs::{BloomFilter, Murmur3};
 use cached::{stores::SizedCache, Cached};
-use rbtree::RBTree;
 use std::sync::Arc;
 use tokio::sync::Mutex;
 use uuid::Uuid;
@@ -48,57 +48,6 @@ pub trait Store {
         &mut self,
         transaction: Transaction<'a, K>,
     ) -> Result<()>;
-}
-
-/// A store backed by a [log-structured merge tree](http://www.benstopford.com/2015/02/14/log-structured-merge-trees)
-pub struct LSMStore {
-    memtable: RBTree<String, Option<Vec<u8>>>,
-    bloom_filter: BloomFilter<Murmur3>,
-}
-
-impl LSMStore {
-    pub fn new() -> Self {
-        Self {
-            memtable: RBTree::new(),
-            // TODO what is the optimal number of items for the bloom filter?
-            bloom_filter: BloomFilter::optimal(Murmur3, 512, 0.01),
-        }
-    }
-
-    // TODO implement memtable flushes to disk
-    // TODO implement segment compaction
-}
-
-#[async_trait]
-impl Store for LSMStore {
-    async fn get<K: AsRef<str> + Send>(&mut self, k: K) -> Result<Option<Vec<u8>>> {
-        if !self.bloom_filter.contains(k.as_ref().as_bytes()) {
-            return Ok(None);
-        }
-        let mem_result = self.memtable.get(&k.as_ref().to_string());
-        // TODO read from disk if value not found in memtable
-        Ok(mem_result.unwrap().as_deref().map(|v| v.to_vec()))
-    }
-
-    async fn transact<'a, K: AsRef<str> + Send>(
-        &mut self,
-        transaction: Transaction<K>,
-    ) -> Result<()> {
-        // TODO write begin_transaction to WAL
-        for instruction in transaction.instructions {
-            match instruction {
-                Set(key, value) => {
-                    self.memtable
-                        .insert(key.as_ref().to_string(), Some(value.to_vec()));
-                    self.bloom_filter.insert(key.as_ref().as_bytes());
-                }
-                Delete(key) => {
-                    self.memtable.insert(key.as_ref().to_string(), None);
-                }
-            };
-        }
-        Ok(())
-    }
 }
 
 /// A basic in memory store for testing
